+++ CHANGELOG.md
@@ -2901,3 +2901,4 @@
 - [2026-07-16 16:37] feat(state): add bi-directional event bus handler for state
 - [2026-07-16 17:12] fix(cache): resolve race condition during cache initialization
 - [2026-07-16 17:28] fix(perf): resolve race condition during perf initialization
+- [2026-07-16 17:49] feat(auth): implement responsive auth view component